==> 8-eLife_revision/Genetic_pairs/2-Plot_phylogeny_data.Rmd <==
---
title: "Description and analysis of genetic pairs phylogenies"
author: "Stefano Giulieri"
date: '28/03/2023'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
setwd(stringr::str_c(here::here(), "/8-eLife_revision/Genetic_pairs"))
msg <- glue::glue("My directory is {getwd()}")
message(msg)
```

Here we plot phylogenies and phylogeny data specifically related to PI contrasting genetic pairs

```{r}
library(tidyverse)
library(ggtree)
library(phytools)
library(ggraph)
library(tidygraph)

rm(list = ls())
```

# Raw data

Pairs with PI data

```{r}
pairs <- readRDS("processed_data/PI_pairs_data/pairs_description_dataframe.Rda")
```

Tree

```{r}
tree <- read.tree("processed_data/tree/iqtree_PI_tips.tree")
ggtree(tree)
```

Phylo data

```{r}
df_ancestral_path <- readRDS("processed_data/PI_pairs_data/get_ancestral_path_dataframe.Rda") 
```

# Merge pairs data with phylo data

```{r}
pairs_with_phylo_data <- pairs %>%
  inner_join(df_ancestral_path) %>%
  arrange(iso1, iso2)
```

Explore association between length of the path and size of the clade

```{r}
pairs_with_phylo_data %>%
  ungroup() %>%
  distinct(path_length, n_offspring, PI_pair) %>%
  ggplot(aes(x = path_length, y = n_offspring)) +
  geom_point() +
  facet_wrap(~PI_pair, scales = "free") +
  theme_bw()

pairs_with_phylo_data %>%
  ungroup() %>%
  distinct(path_length, n_offspring, PI_pair) %>%
  ggplot(aes(x = path_length, y = log10(n_offspring))) +
  geom_point() +
  facet_wrap(~PI_pair, scales = "free") +
  theme_bw()

pairs_with_phylo_data %>%
  ungroup() %>%
  distinct(path_length, n_offspring, PI_pair) %>%
  ggplot(aes(x = path_length, y = log10(n_offspring))) +
  geom_point() +
  facet_wrap(~PI_pair) +
  theme_bw()

pairs_with_phylo_data %>%
  ungroup() %>%
  distinct(path_length, n_offspring, PI_pair) %>%
  ggplot(aes(x = path_length, y = n_offspring)) +
  geom_point() +
  facet_wrap(~PI_pair) +
  scale_y_continuous(trans = "sqrt") +
  theme_bw()

pairs_with_phylo_data %>%
  ungroup() %>%
  distinct(path_length, n_offspring, PI_pair) %>%
  ggplot(aes(x = path_length, y = n_offspring, colour = PI_pair)) +
  geom_point() +
  scale_colour_manual(values = c("red", "blue")) +
  theme_bw()

pairs_with_phylo_data %>%
  ungroup() %>%
  distinct(path_length, n_offspring, PI_pair) %>%
  ggplot(aes(x = path_length, y = log10(n_offspring), colour = PI_pair)) +
  geom_point(alpha = .4, size = 2) +
  scale_colour_manual(values = c("red", "blue")) +
  theme_bw()

pairs_with_phylo_data %>%
  ungroup() %>%
  distinct(path_length, n_offspring, PI_pair) %>%
  ggplot(aes(x = path_length, fill = PI_pair, colour = PI_pair)) +
  geom_density(alpha = .3) +
  scale_fill_manual(values = c("red", "blue")) +
  scale_colour_manual(values = c("red", "blue")) +
  theme_bw()

pairs_with_phylo_data %>%
  ungroup() %>%
  distinct(path_length, n_offspring, PI_pair) %>%
  ggplot(aes(x = n_offspring, fill = PI_pair, colour = PI_pair)) +
  geom_density(alpha = .3) +
  scale_fill_manual(values = c("red", "blue")) +
  scale_colour_manual(values = c("red", "blue")) +
  scale_x_continuous(trans = "sqrt") +
  theme_bw()
```

Compute the heights about the root of each node. This uses `phytools::nodeheight()` function, however, we can extract the same information from the `ggtree`object.

```{r}
# df_node_h <- as_tibble(tree) %>%
#   rowwise() %>%
#   mutate(node_h = phytools::nodeheight(tree, node)) %>%
#   select(node, node_h)
# 
# pairs_with_phylo_data <- pairs_with_phylo_data %>%
#   inner_join(df_node_h, by = c("mrca" = "node"))

# pairs_with_phylo_data %>%
#   ungroup() %>%
#   distinct(node_h, PI_pair) %>%
#   ggplot(aes(x = node_h, fill = PI_pair, colour = PI_pair)) +
#   geom_density(alpha = .3) +
#   scale_fill_manual(values = c("red", "blue")) +
#   scale_colour_manual(values = c("red", "blue")) +
#   # scale_x_continuous(trans = "sqrt") +
#   theme_bw()
```

What about the position of internal nodes on the tree

```{r}
tree_data <- ggtree(tree, branch.length = "none")$data

tree_data %>%
  filter(!isTip) %>%
  ggplot(aes(x = x, y = y)) +
  geom_point() +
  theme_bw()

pairs_with_phylo_data %>%
  inner_join(tree_data, by = c("mrca" = "node")) %>%
  ungroup() %>%
  distinct(x, y, PI_pair) %>%
  ggplot(aes(x = x, y = y)) +
  geom_point() +
  facet_wrap(~PI_pair) +
  theme_bw()

tree_data <- ggtree(tree)$data

tree_data %>%
  filter(!isTip) %>%
  ggplot(aes(x = x, y = y)) +
  geom_point() +
  theme_bw()

pairs_with_phylo_data %>%
  inner_join(tree_data, by = c("mrca" = "node")) %>%
  ungroup() %>%
  distinct(x, y, PI_pair) %>%
  ggplot(aes(x = x, y = y)) +
  geom_point() +
  facet_wrap(~PI_pair) +
  theme_bw()
  
```

Ratio node height/tips height

```{r}
tree_data_min <- tree_data %>%
  select(node, label, x)

df_h_ratio <- pairs_with_phylo_data %>%
  inner_join(tree_data_min %>% select(node, x), by = c("mrca" = "node")) %>%
  rename(x_mrca = x) %>%
  inner_join(tree_data_min %>% select(label, x), by = c("iso1" = "label")) %>%
  rename(x_iso1 = x) %>%
  inner_join(tree_data_min %>% select(label, x), by = c("iso2" = "label")) %>%
  rename(x_iso2 = x) %>%
  ungroup() %>%
  # select(starts_with("x"), PI_pair) %>%
  # distinct() %>%
  rowwise() %>%
  mutate(h_ratio = x_mrca/mean(x_iso1, x_iso2)) 

df_h_ratio %>%
  ggplot(aes(x = h_ratio, fill = PI_pair, colour = PI_pair)) +
  geom_density(alpha = .3) +
  # geom_histogram() +
  scale_fill_manual(values = c("red", "blue")) +
  scale_colour_manual(values = c("red", "blue")) +
  facet_wrap(~PI_pair, scales = "free") +
  theme_bw()

df_h_ratio %>%
  ggplot(aes(x = h_ratio, y = n_offspring, colour = PI_pair)) +
  geom_point() +
  # geom_histogram() +
  scale_colour_manual(values = c("red", "blue")) +
  facet_wrap(~PI_pair, scales = "free") +
  theme_bw()
```

Plot PI pairs

```{r}
pairs_large_clade <- df_h_ratio %>%
  filter(PI_pair & n_offspring > 10) %>%
  ungroup() %>%
  mutate(offset_t = row_number()/50,
         offset = offset_t/10)

x <- pairs_large_clade %>%
  ungroup() %>%
  slice_max(n_offspring, n = 1)

ggtree(tree) +
  geom_cladelab(node = x$mrca, label = x$pair_id) +
  xlim(0,.3) +
  theme_bw()

ggtree(tree) +
  geom_cladelab(data = pairs_large_clade, mapping = aes(node = mrca, label = pair_id, offset.text = offset_t, offset = offset)) +
  xlim(0,.3) 

```

# Save processed data

```{r}
df_h_ratio %>%
  saveRDS("processed_data/PI_pairs_data/pairs_description_with_phylo_dataframe.Rda")
```
